requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
streamlit
jupyterlab
ipykernel
matplotlib
seaborn
plotly
ipython-sql